Move blob tests under subdirectory
