+++ packages/runtime/container-runtime/src/test/blobs/blobHandles.spec.ts
@@ -11,11 +11,11 @@ import { Deferred } from "@fluidframework/core-utils/internal";
 import type { IRuntimeStorageService } from "@fluidframework/runtime-definitions/internal";
 import { createChildLogger } from "@fluidframework/telemetry-utils/internal";
 
-import { BlobManager, type IBlobManagerRuntime } from "../blobManager/index.js";
+import { BlobManager, type IBlobManagerRuntime } from "../../blobManager/index.js";
 import {
 	ContainerFluidHandleContext,
 	type IContainerHandleContextRuntime,
-} from "../containerHandleContext.js";
+} from "../../containerHandleContext.js";
 
 export const failProxy = <T extends object>(handler: Partial<T> = {}): T => {
 	const proxy: T = new Proxy<T>(handler as T, {
